requirements.txt
disnake
rapptz-asqlite
pillow
disnake-jishaku
akinator.py